Add files via upload
The student result management system is similar to a university result portal, It has two sections one is the admin section and the other one is the student section.
The admin can add the students and their results to their end. It would require Username and password. and the student can view their result in the result section once it is uploaded. the student just has to fill their roll number , if the roll number matches with the database it shows the result whereas if it fails to match the rollnumber it throws an error..